updated output for oracle

+++ nibio_inference/clear_cache.py
@@ -9,8 +9,6 @@ args = parser.parse_args()
 with open(args.eval_yaml) as f:
     data = yaml.load(f, Loader=yaml.FullLoader)
 
-print(data['checkpoint_dir'])
-
 # go to checkpoint_dir and '.hydra' folder and open overrides.yaml and find data.dataroot there
 
 with open(os.path.join(data['checkpoint_dir'], '.hydra/overrides.yaml')) as f:
@@ -19,16 +17,20 @@ with open(os.path.join(data['checkpoint_dir'], '.hydra/overrides.yaml')) as f:
 for item in data:
     if 'data.dataroot' in item:
         path = item.split('=')[1].strip()
-        print(path)
-
+      
 # add processed_0.2_test to the path
 path = os.path.join(path, 'treeinsfused')
 path = os.path.join(path, 'processed_0.2_test')
-print(path)
 
-# clear all the files in the directory
-for filename in os.listdir(path):
-    filepath = os.path.join(path, filename)
-    if os.path.isfile(filepath):
-        os.remove(filepath)
-        print(f"Removed: {filepath}")
+print(f"Clearing: {path}")
+
+# clear all the files in the directory if the path exists
+if os.path.exists(path):
+    for filename in os.listdir(path):
+        filepath = os.path.join(path, filename)
+        if os.path.isfile(filepath):
+            os.remove(filepath)
+            print(f"Removed: {filepath}")
+
+else:
+    print(f"Path does not exist: {path}")
